requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
pypdf>=3.17.0
scikit-learn>=1.3.0

# Evaluation Metrics
rouge-score>=0.1.2
nltk>=3.8.1
evaluate>=0.4.0

# API & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
pyngrok>=7.0.0
pydantic>=2.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Jupyter Support
ipywidgets>=8.1.0
notebook>=7.0.0
